chore(buffers): refactor to be able to reuse it
Those changes are there to allow to reuse almost everything from this component in a future one showing a
window instead of a buffer.
To take advantage of what existing I considered that in this case a window will be treated as a buffer with
the extra information of the window it belongs to. So that we can differentiate the current window when the
same buffer is open in multiple windows.