feat(windows): add windows component
Almost identical to the buffers component, it just list the windows of the current tab instead of all
available buffers.
It also check if the buffer shown in the tab bar belongs to the current window or not to avoid having all the
bars for a similar buffer shown as active.