refactor(windows): don't hide help windows
Help buffers are unlisted, so they were hidden.
I think it was a bad idea from the start, I actually tried to hide any kind of windows a plugin could open.
Now I'm starting to think it's better to show any visible window in the "base layout" and hide the floating
windows.