refactor(windows): create only from winnr and add default
This way the interface is cleaner since identical for `Buffers` and `Windows`.
I forgot that when listing the buffers if the current buffer is not part of the one we usually show then we
generate a new object on the fly.
This is why I still needed to handle the `bufnr` manually in the `Windows::new_buffer` function.