Add typings for LSP
With these changes, `lua-language-server` will show type hints for the
config and all its fields.

This is strongly based on the typings provided by @zorgick in #563. I
omitted the defaults. When putting the typing defitions together with
the declaration of the default instance, the LSP will use its values as
default values.

Closes: https://github.com/nvim-lualine/lualine.nvim/issues/563
Co-Authored-By: @zorgick