feat: allow custom theming for select mode
All select modes where linked hard to use the visual theme which makes
adjustments impossible. Now, they are linked to their own suffix and
thus theming is possible. To avoid falling back to normal colors in case
no definition of select is found, it is linked to use the colors defined
for visual (basically, restoring the old behavior).