Fix #796 and close #797
If there are errors loading nvim-base16, the setup_base16() call will
throw an error instead of returning defaults. This has the effect of
falling back to the 'auto' theme, allowing the colorscheme to remain
more consistent with the existing theme as opposed to using a default
that doesn't match the exising theme.