Add ability to format tabs with highlights
This adds a handling of optional `fmt` return value that is a table of
`{highlight = "*", name = "*"}` or `{name = "*"}` tables. The primary
goal is to have highlights not count to the visible character limit.
Without it, attempts to format the tab name with highlights cause nearly
immediate ellipsification.

Due to crossbleed in the test setup, a buffers component test was also
updated to test through _buffer_ commands. Frankly, when using
bufferline, one should forget about tab commands, as the real state of
vim windows is hidden and unpredictable.

Mode handling had been forced to not interact with table processing at
all. If user provides a detailed table, they are assumed to be in
control.