feat: native support for mini.icons
## Details

`mini.icons` is an alternative to `nvim-web-devicons` that is getting
support from quite a few plugins. This change adds native support for
the plugin so users do not need to mock it via `mock_nvim_web_devicons`.

Currently `nvim-web-devicons` is used in the `buffer` and `filetype`
components. Created a separate `icons` module to store the icon
related logic which now gets used by both of these components.

The logic is to first check if `mini.icons` exists and has been setup,
if so for the `buffer` component use `get` with the `file` category, for
`filetype` use `get` with the `filetype` category. It may make sense to
use the `file` category for both.

From there the fallback to use `nvim-web-devicons` or `vim-devicons` is
kept identical.